feat(api): add Prisma client for database management and enhance login redirection based on user roles
