+++ apps/api/src/db.js
@@ -1,0 +1,6 @@
+const { PrismaClient } = require("@prisma/client")
+
+const prisma = new PrismaClient()
+
+module.exports = prisma
+
